requirements.txt
colorama==0.4.6
dnspython==2.4.2
npyscreen==4.10.5
pymongo==3.11.0
python-dotenv==1.0.0
six==1.16.0
windows-curses==2.3.1
